Removed unused `database` field from SugarRecord
This also produced a NullPointerException upon call to the constructor if the SugarApp was not added in the manifest.

Now the NullPointerException is produced upon save of the object, which makes more sense in my opinion, although it would be preferable to have another type of exception being produced given that particular case.