Added ObjectCache support in SugarRecord.
Object cache allows that every instance of a database record as an object references the same object.
This allows multiple threads to have a shared instance of objects.

ObjectCache is implemented using weak reference, therefore when the object is no longer referenced directly
the GC will release it.

The object cache is disabled by default and can be enabled by using OBJECT_CACHE true in the manifest.