Make SugarRecord API more versatile
this should not change the API, but makes it easier to use.
Especially for cases like wanting to count for one specific value,
resulting in a where with only one or a few args.
Currently we need to create a String array for that, resulting in a
ugly new String[]{"bla"} block.
Now it is possible to directly pass "bla"

count was the motivation for this change, but I changed other signatures
where this was possible as well.